e2e_tests/tests/requirements.txt
appdirs
# pytest 6.0 has linter-breaking changes
pytest>=6.0.1
pytest-timeout
torch==1.11.0
torchvision==0.12.0
tensorflow==2.12.0; sys_platform != 'darwin' or platform_machine != 'arm64'
tensorflow-macos==2.11.0; sys_platform == 'darwin' and platform_machine == 'arm64'
pandas
pyyaml
docker
python-dateutil
kubernetes
# numby>1.20 and h5py==2.10 are incompatible, force newer versions of
# of both to avoid AttributeError: module 'numpy' has no attribute 'typeDict'
# when pytest executes e2e_tests.
numpy>=1.20
h5py>=3
ray[default,tune]
pyarrow
# Pydantic V2 has changes that break existing ray tests
pydantic<2
